+++ constants/constants.go
@@ -34,9 +34,9 @@ const (
 	EnvConfig                    = "DEBUG_CONFIG"
 	EnvDebugAws                  = "DEBUG_AWS"
 	EnvLogDebug                  = "LOG_DEBUG"
+	EnvLogColor                  = "LOG_COLOR"
 	EnvStackCreation             = "STACK_CREATION_TIMEOUT"
 	EnvStackCreationPollInterval = "STACK_CREATION_POLL_INTERVAL"
-	EnvNoLogColor                = "NO_LOG_COLOR"
 	EnvDevMode                   = "DEV_MODE"
 
 	// Registry-based deployment