+++ daemon/elb_registration/elb_registration.go
@@ -20,12 +20,14 @@ import (
 	"github.com/adobe-platform/porter/constants"
 	"github.com/adobe-platform/porter/daemon/identity"
 	"github.com/adobe-platform/porter/logger"
+	"github.com/adobe-platform/porter/util"
 	elblib "github.com/aws/aws-sdk-go/service/elb"
-	"github.com/inconshreveable/log15"
 )
 
 func Call() {
-	log := logger.Daemon("AWS_STACKID", os.Getenv("AWS_STACKID"))
+	stackId := os.Getenv("AWS_STACKID")
+	log := logger.Daemon("AWS_STACKID", stackId)
+
 	log.Info("Auto ELB instance registration")
 
 	elbCSV := os.Getenv("ELBS")
@@ -38,6 +40,7 @@ func Call() {
 	if err != nil {
 		return
 	}
+	instanceIds := []string{ii.Instance.InstanceID}
 
 	elbClient := elb.New(aws_session.Get(ii.AwsCreds.Region))
 
@@ -45,45 +48,58 @@ func Call() {
 
 outer:
 	for _, elbName := range elbNames {
-		log = log.New("LoadBalancerName", elbName)
+		log := log.New("LoadBalancerName", elbName)
 
-		tagDescriptions, err := elb.DescribeTags(elbClient, elbName)
-		if err != nil {
-			log.Error("elb.DescribeTags", "Error", err)
-			log.Warn("instance autoregistration is broken")
+		var tagDescriptions []*elblib.TagDescription
+		var err error
+
+		retryMsg := func(i int) { log.Warn("elb.DescribeTags retrying", "Count", i) }
+		if !util.SuccessRetryer(8, retryMsg, func() bool {
+
+			tagDescriptions, err = elb.DescribeTags(elbClient, elbName)
+			if err != nil {
+				log.Error("elb.DescribeTags", "Error", err)
+				return false
+			}
+
+			return true
+		}) {
+			log.Warn("elb.DescribeTags failed")
 			continue
 		}
 
-		for _, tagDescription := range tagDescriptions {
+		for _, tagDescription := range tagDescriptions{
 			for _, tag := range tagDescription.Tags {
-				if tag.Key != nil && *tag.Key == constants.PorterStackIdTag {
-					if os.Getenv("AWS_STACKID") == *tag.Value {
-						log.Info("Instance found its stack tagged to the ELB")
-						go tryRegistration(log, elbClient, elbName)
-					} else {
-						log.Info("Instance did not find its stack tagged to the ELB")
-					}
+				if tag.Key == nil || *tag.Key != constants.PorterStackIdTag {
+					continue
+				}
+
+				if *tag.Value != stackId {
+					log.Info("Instance is NOT associated with a stack that was promoted into this ELB")
 					continue outer
 				}
+
+				log.Info("Instance IS associated with a stack that was promoted into this ELB")
+				log.Info("RegisterInstancesWithLoadBalancer", "InstanceId", ii.Instance.InstanceID)
+
+				retryMsg := func(i int) { log.Warn("elb.RegisterInstancesWithLoadBalancer retrying", "Count", i) }
+				if !util.SuccessRetryer(8, retryMsg, func() bool {
+
+					_, err = elb.RegisterInstancesWithLoadBalancer(elbClient, elbName, instanceIds)
+					if err != nil {
+						log.Error("RegisterInstancesWithLoadBalancer", "Error", err)
+						return false
+					}
+
+					return true
+				}) {
+					log.Error("Instance Registration failed")
+				}
+
+				continue outer
 			}
 		}
 
 		log.Warn("Didn't find tag key " + constants.PorterStackIdTag)
-	}
-}
-
-func tryRegistration(log log15.Logger, elbClient *elblib.ELB, elbName string) {
-	ii, err := identity.Get(log)
-	if err != nil {
-		return
-	}
-
-	instanceIds := []string{ii.Instance.InstanceID}
-
-	log.Info("RegisterInstancesWithLoadBalancer", "InstanceId", ii.Instance.InstanceID)
-	_, err = elb.RegisterInstancesWithLoadBalancer(elbClient, elbName, instanceIds)
-	if err != nil {
-		log.Error("RegisterInstancesWithLoadBalancer", "Error", err)
-		log.Error("instance autoregistration failed")
 	}
 }